Fix for fwd and for exec
fwd: when a value is forwarded as FWD_ALU it will also
be presented to the fwd unit as FWD_WB later. But then
there might be a newer version of the register forwarded
as FWD_ALU, so FWD_ALU needs to be of higher priority.

exec: forwardA, forwardB, mem_aluresult, wb_result were
missing from the process sensitivity list and so the
forwarding didn't work